requirements.txt
alpaca-trade-api==3.*
pandas
requests
python-dotenv
feedparser
python-dotenv
